Move all templates to plugins

+++ src/main/kotlin/io/guthix/oldscape/server/world/entity/interest/VarpManager.kt
@@ -17,7 +17,7 @@ package io.guthix.oldscape.server.world.entity.interest
 
 import io.guthix.oldscape.server.net.game.out.VarpLargePacket
 import io.guthix.oldscape.server.net.game.out.VarpSmallPacket
-import io.guthix.oldscape.server.template.api.VarbitTemplates
+import io.guthix.oldscape.server.content.VarbitTemplates
 import io.guthix.oldscape.server.world.World
 import io.guthix.oldscape.server.world.entity.Player
 import io.netty.channel.ChannelFuture
@@ -35,11 +35,11 @@ class VarpManager : InterestManager {
 
     fun updateVarbit(id: Int, value: Int) {
         val config = VarbitTemplates[id]
-        val bitSize = (config.msb.toInt() - config.lsb.toInt()) + 1
+        val bitSize = (config.msb- config.lsb) + 1
         if (value > 2.0.pow(bitSize) - 1) throw IllegalArgumentException("Value $value to big for this varbit.")
         var curVarp = varps[config.varpId] ?: 0
-        curVarp = curVarp.clearBits(config.msb.toInt(), config.lsb.toInt())
-        curVarp = curVarp or value shl config.lsb.toInt()
+        curVarp = curVarp.clearBits(config.msb, config.lsb)
+        curVarp = curVarp or value shl config.lsb
         varps[config.varpId] = curVarp
         changes[config.varpId] = curVarp
     }
